+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Adding an older version of a package to a Pacman repo now properly reports an exception instead of 
   silently doing nothing.
 - `AptRepo` now leaves any non `*.deb` files alone in a destination pool (as was promised by the docs).
+- `gpg` passphrase is no longer passed on the command line. 
 - Dotless signature files are now handled correctly on the command line.
 - Typos and grammar in documentation.
 - Tests now work again.